requirements.txt
mcp
python-dotenv
openai>=1.12.0
pyautogen[teachable]==0.2.9
requests>=2.31.0
flaml[automl]>=1.8.0
python-dotenv>=0.19.0
gitpython
azure-ai-inference
autogen-agentchat
autogen-ext[openai,azure,mcp]